stock-data-with-yfinance/requirements.txt
yfinance
plotly.express